wip: nextstop agent
