+++ parrot/handlers/agents/__init__.py
@@ -1,0 +1,1 @@
+from .abstract import AgentHandler
